requirements.txt
Django==2.2.10 
djangorestframework==3.9.1
bcrypt==3.1.7 
psycopg2==2.8.4
gunicorn == 20.0.4
django-heroku==0.3.1